SSE-5 Provide Heartbeat signal out of SSE
- Adds full set of Heartbeat Endpoint and Service including a thread
that generates "Heartbeat" messages on a 10-second interval.

NOTE: This is distinct from the internal KeepAlive that is used
underneath the covers to keep the connection open.

- Refactors shared code between Game State and Heartbeat into 'common'
package.
- Prepares (but not completes) work for adding a tether message. This
does take into account the presence or absence of an Outing ID -- no
tether unless you're part of an Outing whose Guide you can tether to.