SSE-20 SSE Error - trying to re-establish a session every second.
- Separates UserChannel from OutingChannel.
- Adds services for UserChannel and OutingChannel; these replace CommonChannelService.
- Continues cleanup of factories that had been used to generate Events of appropriate types.
- KeepAlive is limited to the SseBroadcaster which is itself limited to a single EventOutput;
each of these are associated one-to-one with a UserChannel.
- Adds UserChannel to an OutingChannel's broadcaster and then removed from that broadcaster
should the UserChannel go stale.

Currently, it takes about three minutes for a stale connection to be recognized as stale and
to be closed. In the meanwhile, clients are able to establish new sessions without worrying
about the state of any previous connections, stale, closed or otherwise.

The side-effect of this behaviour is that it won't be possible for a single user to use the
same account to participate in Clue Ride via two different devices, browser tabs, or any other
session. The last account to start a session for a given user ID will replace any other existing
connections and render them inoperative.